third_party/WebKit/LayoutTests/fast/dom/shadow/querySelector-for-multiple-shadowroots-expected.txt

CONSOLE WARNING: line 87: Calling Element.createShadowRoot() for an element which already hosts a shadow root is deprecated and will be removed in M63, around December 2017. See https://www.chromestatus.com/features/4668884095336448 for more details.
crbug.com/337616: test for querySelectorAll with ::shadow and /deep/

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll("#foo-host::shadow span").length is 6
PASS document.querySelectorAll("#foo-host::shadow span")[0].id is "not-top1"
PASS document.querySelectorAll("#foo-host::shadow span")[1].id is "top1"
PASS document.querySelectorAll("#foo-host::shadow span")[2].id is "not-top2"
PASS document.querySelectorAll("#foo-host::shadow span")[3].id is "top2"
PASS document.querySelectorAll("#foo-host::shadow span")[4].id is "not-top3"
PASS document.querySelectorAll("#foo-host::shadow span")[5].id is "top3"
PASS shadowRoot1.querySelectorAll(":host::shadow span").length is 2
PASS shadowRoot1.querySelectorAll(":host::shadow span")[0].id is "not-top1"
PASS shadowRoot1.querySelectorAll(":host::shadow span")[1].id is "top1"
PASS shadowRoot2.querySelectorAll(":host::shadow span").length is 2
PASS shadowRoot2.querySelectorAll(":host::shadow span")[0].id is "not-top2"
PASS shadowRoot2.querySelectorAll(":host::shadow span")[1].id is "top2"
PASS shadowRoot3.querySelectorAll(":host::shadow span").length is 2
PASS shadowRoot3.querySelectorAll(":host::shadow span")[0].id is "not-top3"
PASS shadowRoot3.querySelectorAll(":host::shadow span")[1].id is "top3"
PASS successfullyParsed is true

TEST COMPLETE
